+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 * Add `SyncMessagePort.receiveMessageIfAvailable()`.
 
+* Add `timeout` and `timeoutValue` options to
+  `SyncMessagePort.receiveMessage()`.
+
 ## 1.0.0
 
 * Initial release